+++ app/Matches.tsx
@@ -12,7 +12,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import { useNavigation } from '@react-navigation/native';
 import { EventProfile, Like, Message } from '../lib/api/entities';
 import { Heart, MessageCircle, Users, Sparkles } from 'lucide-react-native';
-import ChatModal from './ChatModal';
+import ChatModal from '../components/modals/ChatModal';
 
 export default function MatchesScreen() {
   const navigation = useNavigation();
