+++ components/modals/EventCodeEntry.tsx
