+++ components/modals/ProfileDetailModal.tsx
