+++ components/modals/QRScanner.tsx
